This commit finalizes a set of changes related to service discovery and clustering.
Specifically, actors can now register themselves as "services" with a shard, so that other
actors can request the name of the service and receive a list of addresses which provide the
service.

The shards, in turn, can register as shards with the messaging system. This allows them to receive
*broadcasts* wich are sent to all known shards. The shards use this, and knowledge of their neighbors
to spread information on local and remote service providers. The main mechanism being a low intensity
whisper daemon in each shard that stochastically selects some number of other shards and whispers its most
up to date information to them. This has been designed so that the information coalesces and floats
towards the edges of the address space, where, with some low probability, actors may trigger a broadcast
in which they spread their information to the entire set of shards. This way, the whole cluster should
receive accurate and up to date service information from the other shards with comparatively low traffic,
growing linearly with the number of shards.

Other improvements include:

- Adding a socksend method to the BackActor class for convenience.
- Making the BackActor use subsession name strings for replacement instead of the factory function itself.
- Adding the capability to attach daemons to actors, that will act in the background as separate tasks.
- Enabling actors to define on_start and on_stop methods for handling asynchronous setup and shutdown routines.
- Changed SimpleMessage so that it always serializes and deserializes messages as JSON, because not doing so
  may introduce obscure bugs, such as forwarding references instead of copies of mutable objects.
- Added a primitve error handling callback for actors so that they announce why they have crashed instead of remaining silent.
- Added loop parameters so that we may specify which event loop to run things on.
- Numerous smaller bug fixes.