A large number of changes have been made in order to support cross-shard spawning. In particular, the following has been changed:
1. The shard synchronization method has been replaced with a new method centered around time stamped information bits collected in buckets.
2. A new naming convention has been instated which follows the format [shard]:[actor] in which the shard actor has the address [shard]:[shard].
3. An entirely new message system called RoutedMessage has been written, which utilizes the aforementioned naming convention.
4. Intershard spawning has been introduced.

These changes were all motivated by the last item in the list.

Shards are allowed to be hetrogenous. For example, some nodes may have specialized
hardware such as graphics cards, storage facilities etc., and some actors may have to reside only on those shards. Therefore, availability of
actor types must be communicated and synchronized across the cluster. The previous synchronization algorithm was not sufficiently flexible to
easily accomodate new types of information. In the new system, all shards maintain a bag of "information", that is, time stamped snippets of
information about things like whether a certain service exists on a certain shard. These bags are updated as new snippets arrive over the messaging system.
Whenever a snippet arrives, it replaces an already present piece of information if it is newer. The shard's internal registries are then kept separately
and are compiled separately from the information present in the bag at the time of the update. This new system easily allows new types of information
to be added, and one such new type of information, availability of factories, was added.

It was found that the old CombinedMessage system spawned new actors too slowly. This was partly due to overhead associated with spawning new queues, and partly due to
publisher confirms. Therefore, a new system was devised in which each shard has one RabbitMQ queue, from which messages are distributed using the ordinary
SimpleMessage system. This made spawning significantly faster, but the system is still far too slow. On a development laptop, locally spawned shards could
spawn at a rate of over 10 000 per second, but even several remote shards were not able to spawn actors at a rate of more than around 1 600 per second. This
slowness also seems to affect other interactions between shards, and seems to be entirely due to inherent performance limitations in RabbitMQ. As a result of
having merely one queue per shard in the RoutedMessage system, we needed to provide the messaging system with a hint so that it would know which of these queues
to deposit a message into. Therefore a new module called "addressing.py" was introduced which aids in the generation of consistent addresses on the new format.

Once these major changes had been introduced, a simple inter-shard spawning mechanism could be added. The shard actor now injects a method into all actors it spawns,
which allows the actor to directly interact with the shard's internal registries when spawning. This allows the actor to generate a suitable name for its new child without
having to negotiate with any shard actors over the messaging system. This is essential, since it allows us to encapsulate spawning into a method on the actor instance
whether it relies on sending a message to a different shard or wether it directly invokes the spawn function on the local shard.

Despite these changes, and the resulting fully functional cross-shard spawning mechanism, we need to evaluate other messaging backends for Skick in order to ensure a sufficient
inter-shard messaging throughput. In the short term, Redis pub/sub should be easy enough to implement, and should have sufficient performance. However, it may have
insufficient guarantees for things like message ordering.